🏗 <Iconed> Change the dynamic svg loading strategy
Use 'raw-loader' to get the .svg contents and stick them in the DOM.
Previously, we expected all SVGs to have a specific viewBox, which is ridiculous.
This way, all SVGs can be imported, and we no longer have to tweak the svg file itself with "fill="var(--fill)"" (as this is done automatically) with a .replace when loading the file.